e2e: move server into playwright (#4397)
