+++ e2e/integration/vanilla-node/e2e/utils.ts
@@ -2,6 +2,7 @@ import type { Page } from "@playwright/test";
 import { type ChildProcessWithoutNullStreams, spawn } from "node:child_process";
 import { createRequire } from "node:module";
 import { fileURLToPath } from "node:url";
+import { createAuthServer } from "./app";
 
 const terminate = createRequire(import.meta.url)(
 	// use terminate instead of cp.kill,
@@ -21,7 +22,7 @@ export async function runClient<R>(
 }
 
 export function setup() {
-	let serverChild: ChildProcessWithoutNullStreams;
+	let server: Awaited<ReturnType<typeof createAuthServer>>;
 	let clientChild: ChildProcessWithoutNullStreams;
 	const ref: {
 		clientPort: number;
@@ -33,21 +34,10 @@ export function setup() {
 	return {
 		ref,
 		start: async () => {
-			serverChild = spawn("pnpm", ["run", "start:server"], {
-				cwd: root,
-				stdio: "pipe",
-			});
+			server = await createAuthServer();
 			clientChild = spawn("pnpm", ["run", "start:client"], {
 				cwd: root,
 				stdio: "pipe",
-			});
-			serverChild.stderr.on("data", (data) => {
-				const message = data.toString();
-				console.log(message);
-			});
-			serverChild.stdout.on("data", (data) => {
-				const message = data.toString();
-				console.log(message);
 			});
 			clientChild.stderr.on("data", (data) => {
 				const message = data.toString();
@@ -59,6 +49,15 @@ export function setup() {
 			});
 
 			await Promise.all([
+				new Promise<void>((resolve) => {
+					server.listen(0, "0.0.0.0", () => {
+						const address = server.address();
+						if (address && typeof address === "object") {
+							ref.serverPort = address.port;
+							resolve();
+						}
+					});
+				}),
 				new Promise<void>((resolve) => {
 					clientChild.stdout.on("data", (data) => {
 						const message = data.toString();
@@ -73,27 +72,11 @@ export function setup() {
 						}
 					});
 				}),
-				new Promise<void>((resolve) => {
-					serverChild.stdout.on("data", (data) => {
-						const message = data.toString();
-						// find: http://localhost:3000
-						if (message.includes("http://localhost:")) {
-							const port: string = message
-								.split("http://localhost:")[1]
-								.split("/")[0]
-								.trim();
-							ref.serverPort = Number(port.replace(/\x1b\[[0-9;]*m/g, ""));
-							resolve();
-						}
-					});
-				}),
 			]);
 		},
 		clean: async () => {
-			await Promise.all([
-				terminate(clientChild.pid!),
-				terminate(serverChild.pid!),
-			]);
+			await terminate(clientChild.pid!);
+			server.close();
 		},
 	};
 }
