+++ e2e/integration/vanilla-node/package.json
@@ -3,12 +3,9 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "start:server": "tsx ./src/server.ts",
     "start:client": "vite"
   },
   "devDependencies": {
-    "tsx": "^4.20.5",
-    "typescript": "^5.9.2",
     "vite": "^7.1.4"
   },
   "dependencies": {
