+++ e2e/integration/vanilla-node/src/main.ts
@@ -1,7 +1,10 @@
 import { createAuthClient } from "better-auth/client";
 
+const search = new URLSearchParams(window.location.search);
+const port = search.get("port");
+
 const client = createAuthClient({
-	baseURL: "http://localhost:3000",
+	baseURL: `http://localhost:${port ?? 3000}`,
 });
 
 declare global {
